test: add unit tests
- Add initial unit testing framework via Vitest
- Reorganize e2e tests in their own folder
- Reorganize package.json scripts to distinguish between e2e tests and unit tests
- Update testing documentation
