+++ package.json
@@ -24,10 +24,11 @@
     "pre-commit": "node ./scripts/pre-commit.js",
     "prebuild": "npm run version",
     "preview": "svelte-kit preview",
-    "test": "playwright test",
-    "test:codegen": "playwright codegen http://localhost:3000",
-    "test:debug": "PWDEBUG=1 playwright test",
-    "test:dev": "AUTH_TYPE=none ORIGIN=http://localhost:3000 node build",
+    "test:e2e": "playwright test",
+    "test:e2e:codegen": "playwright codegen http://localhost:3000",
+    "test:e2e:debug": "PWDEBUG=1 playwright test",
+    "test:e2e:dev": "AUTH_TYPE=none ORIGIN=http://localhost:3000 node build",
+    "test:unit": "vitest run",
     "version": "node ./scripts/version.js"
   },
   "devDependencies": {
@@ -35,6 +36,7 @@
     "@playwright/test": "^1.22.2",
     "@sveltejs/adapter-node": "1.0.0-next.78",
     "@sveltejs/kit": "1.0.0-next.350",
+    "@testing-library/svelte": "^3.1.3",
     "@types/cookie": "^0.5.1",
     "@types/d3-axis": "^3.0.1",
     "@types/d3-brush": "^3.0.1",
@@ -63,6 +65,7 @@
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-svelte3": "^4.0.0",
     "graphql-ws": "^5.9.0",
+    "jsdom": "^20.0.0",
     "lodash-es": "^4.17.21",
     "modern-css-reset": "^1.4.0",
     "monaco-editor": "^0.33.0",
@@ -78,6 +81,8 @@
     "toastify-js": "^1.11.2",
     "tslib": "^2.4.0",
     "typescript": "^4.7.4",
-    "unique-names-generator": "^4.7.1"
+    "unique-names-generator": "^4.7.1",
+    "vitest": "^0.15.2",
+    "vitest-svelte-kit": "^0.0.6"
   }
 }
