+++ tests/e2e/data/command-dictionary.xml
