+++ tsconfig.json
@@ -16,5 +16,5 @@
     "target": "es2020"
   },
   "extends": "./.svelte-kit/tsconfig.json",
-  "include": ["src/**/*.d.ts", "src/**/*.js", "src/**/*.ts", "src/**/*.svelte"]
+  "include": ["src/**/*.d.ts", "src/**/*.js", "src/**/*.ts", "src/**/*.svelte", "tests/unit/**/*.test.ts"]
 }
