+++ vitest.config.js
@@ -1,0 +1,3 @@
+import { extractFromSvelteConfig } from 'vitest-svelte-kit';
+
+export default extractFromSvelteConfig();
